:sparkles: Feat: add module Reinhardt
