--- src/types/hero-name.type.ts
@@ -1,4 +1,0 @@
-export enum HeroName {
-  Ana = 'Ana',
-  Cassidy = 'Cassidy',
-}
