--- src/types/role.type.ts
@@ -1,5 +1,0 @@
-export enum Role {
-  Damage = 'damage',
-  Tank = 'tank',
-  Support = 'support',
-}
